pyproject.toml
[project]
name = "redshift-mcp-server"
version = "1.0.0"
description = "Redshift MCP Server for AI agents to interact with Amazon Redshift"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.0.0",
    "redshift-connector>=2.1.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.0",
    "sqlalchemy>=2.0.0",
]

[project.scripts]
redshift-mcp = "redshift_mcp_server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
